src/updater.py
```python
import requests
from bs4 import BeautifulSoup
from src.logger import Logger
from src.error_handler import ErrorHandler

class Updater:
    def __init__(self):
        self.logger = Logger()
        self.error_handler = ErrorHandler()

    def check_for_updates(self, url):
        try:
            response = requests.get(url)
            soup = BeautifulSoup(response.text, 'html.parser')

            # Check if the website structure has changed
            if not self.is_structure_same(soup):
                self.logger.log("Website structure has changed. Updating the scraper...")
                self.update_scraper()
            else:
                self.logger.log("No updates needed.")

        except Exception as e:
            self.error_handler.handle_error(e)

    def is_structure_same(self, soup):
        # This function should be implemented based on the specific structure of the websites being scraped
        # For example, it can check if certain key elements are still present in the HTML
        pass

    def update_scraper(self):
        # This function should be implemented to update the scraper based on the new structure of the websites
        # This could involve changing the way the HTML is parsed, the way data is extracted, etc.
        pass
```