tests/test_updater.py
```python
import unittest
from src import updater

class TestUpdater(unittest.TestCase):

    def setUp(self):
        self.updater = updater.Updater()

    def test_check_for_updates(self):
        result = self.updater.check_for_updates()
        self.assertIsInstance(result, bool)

    def test_download_updates(self):
        result = self.updater.download_updates()
        self.assertIsInstance(result, bool)

    def test_apply_updates(self):
        result = self.updater.apply_updates()
        self.assertIsInstance(result, bool)

    def test_rollback_updates(self):
        result = self.updater.rollback_updates()
        self.assertIsInstance(result, bool)

if __name__ == '__main__':
    unittest.main()
```